--- packages/website/src/pages/play/counter/html.js
@@ -1,17 +1,0 @@
-export default `<!-- playground-fold -->
-<!DOCTYPE html>
-<head>
-</head>
-
-<body style="background-color: var(--sapBackgroundColor); color: var(--sapTextColor);">
-  <!-- playground-fold-end -->
-
-  <my-counter>
-    <h1>UI5 Web Components</h1>
-  </my-counter>
-  <!-- playground-fold -->
-  <script type="module" src="main.js"></script>
-</body>
-
-<!-- playground-fold-end -->
-`;
