--- packages/website/src/pages/play/hello-world/html.js
@@ -1,16 +1,0 @@
-export default `<!-- playground-fold -->
-<!DOCTYPE html>
-<head>
-</head>
-
-<body style="background-color: var(--sapBackgroundColor); color: var(--sapTextColor);">
-  <!-- playground-fold-end -->
-
-  <my-element></my-element>
-  <my-element name="UI5 Web Components"></my-element>
-  <!-- playground-fold -->
-  <script type="module" src="main.js"></script>
-</body>
-
-<!-- playground-fold-end -->
-`;
